+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Imp/CertsSpec.hs
@@ -12,9 +12,8 @@ module Test.Cardano.Ledger.Conway.Imp.CertsSpec (conwayOnlySpec, spec) where
 
 import Cardano.Ledger.BaseTypes (EpochInterval (..), Mismatch (..))
 import Cardano.Ledger.Coin (Coin (..))
-import Cardano.Ledger.Conway (hardforkConwayMoveWithdrawalsAndDRepChecksToLedgerRule)
 import Cardano.Ledger.Conway.Core
-import Cardano.Ledger.Conway.Rules (ConwayCertsPredFailure (..), ConwayLedgerPredFailure (..))
+import Cardano.Ledger.Conway.Rules (ConwayLedgerPredFailure (..))
 import Cardano.Ledger.Credential (Credential (..))
 import Cardano.Ledger.DRep (DRep (..))
 import Cardano.Ledger.Plutus (SLanguage (SPlutusV3), hashPlutusScript)
@@ -34,7 +33,6 @@ conwayOnlySpec = describe "CERTS" $ do
   describe "Withdrawals" $ do
     it "Withdrawing from an unregistered staking address" $ do
       modifyPParams $ ppGovActionLifetimeL .~ EpochInterval 2
-      pv <- getsPParams @era ppProtocolVersionL
 
       stakeKey <- freshKeyHash
       accountAddress <- getAccountAddressFor $ KeyHashObj stakeKey
@@ -45,11 +43,8 @@ conwayOnlySpec = describe "CERTS" $ do
               & withdrawalsTxBodyL
                 .~ Withdrawals [(accountAddress, Coin 20)]
         notInRewardsFailure =
-          ( if hardforkConwayMoveWithdrawalsAndDRepChecksToLedgerRule pv
-              then injectFailure . ConwayWithdrawalsMissingAccounts @era
-              else injectFailure . WithdrawalsNotInRewardsCERTS @era
-          )
-            $ Withdrawals [(accountAddress, Coin 20)]
+          (injectFailure . ConwayWithdrawalsMissingAccounts @era) $
+            Withdrawals [(accountAddress, Coin 20)]
        in
         submitBootstrapAware
           (submitTx_ tx)
@@ -72,11 +67,8 @@ conwayOnlySpec = describe "CERTS" $ do
               & withdrawalsTxBodyL
                 .~ Withdrawals [(accountAddress, zero), (registeredAccountAddress, reward)]
         notInRewardsFailure =
-          ( if hardforkConwayMoveWithdrawalsAndDRepChecksToLedgerRule pv
-              then injectFailure . ConwayWithdrawalsMissingAccounts @era
-              else injectFailure . WithdrawalsNotInRewardsCERTS @era
-          )
-            $ Withdrawals [(accountAddress, zero)]
+          (injectFailure . ConwayWithdrawalsMissingAccounts @era) $
+            Withdrawals [(accountAddress, zero)]
        in
         submitBootstrapAware
           (submitTx_ tx)
@@ -99,7 +91,6 @@ spec = describe "CERTS" $ do
   describe "Withdrawals" $ do
     it "Withdrawing the wrong amount" $ do
       modifyPParams $ ppGovActionLifetimeL .~ EpochInterval 2
-      pv <- getsPParams @era ppProtocolVersionL
 
       (accountAddress1, reward1, stakeKey1) <- setupAccountAddress
       (accountAddress2, reward2, stakeKey2) <- setupAccountAddress
@@ -120,14 +111,9 @@ spec = describe "CERTS" $ do
                   , (accountAddress2, reward2)
                   ]
         )
-        [ if hardforkConwayMoveWithdrawalsAndDRepChecksToLedgerRule pv
-            then
-              injectFailure . ConwayIncompleteWithdrawals @era $
-                NEM.singleton accountAddress1 $
-                  Mismatch (reward1 <+> Coin 1) reward1
-            else
-              injectFailure . WithdrawalsNotInRewardsCERTS @era $
-                Withdrawals [(accountAddress1, reward1 <+> Coin 1)]
+        [ injectFailure . ConwayIncompleteWithdrawals @era $
+            NEM.singleton accountAddress1 $
+              Mismatch (reward1 <+> Coin 1) reward1
         ]
 
       submitFailingSubsetTx
@@ -139,14 +125,9 @@ spec = describe "CERTS" $ do
                 .~ Withdrawals
                   [(accountAddress1, zero)]
         )
-        [ if hardforkConwayMoveWithdrawalsAndDRepChecksToLedgerRule pv
-            then
-              injectFailure . ConwayIncompleteWithdrawals @era $
-                NEM.singleton accountAddress1 $
-                  Mismatch zero reward1
-            else
-              injectFailure . WithdrawalsNotInRewardsCERTS @era $
-                Withdrawals [(accountAddress1, zero)]
+        [ injectFailure . ConwayIncompleteWithdrawals @era $
+            NEM.singleton accountAddress1 $
+              Mismatch zero reward1
         ]
 
 setupAccountAddress ::